+++ src/types.rs
@@ -1,10 +1,10 @@
-pub type WORD  = u32;
+pub type WORD = u32;
 pub type HWORD = u16;
 pub type BYTE = u8;
 
-pub type REGISTER = usize;
+pub type REGISTER = u32;
 
-pub trait ARMBITS {} 
+pub trait ARMBITS {}
 impl ARMBITS for WORD {}
 impl ARMBITS for HWORD {}
 impl ARMBITS for BYTE {}
